+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for calldata variables ([#33](https://github.com/edb-rs/edb/pull/33))
 - Add new `r` and `R` commands in code panel (`run`/`runback` in terminal panels) to run forward/backward until the next breakpoint
+- Add `edb server` which collectively spawns edb debug server. ([#46](https://github.com/edb-rs/edb/pull/46))
 
 ### Fixed